🧹 rename default reporter to cli reporter since its used for multiple formats
